👍 getWeekRange関数をutilsに移動させ、それを利用して月別→週別の範囲も引き継げるようにした
