+++ packages/shared/src/components/parts/range-selector.component.tsx
@@ -7,8 +7,8 @@ import {
   PopoverContent,
   PopoverTrigger,
 } from "@fukui-kanko/shared/components/ui";
+import { formatDate, getMaxDate, getMinDate, getWeekRange } from "@fukui-kanko/shared/utils";
 import { CalendarIcon } from "@primer/octicons-react";
-import { FIRST_WEEK_END_DATE, formatDate, getMaxDate, getMinDate } from "../../utils/utils";
 
 type WeekRange = { from: Date; to: Date } | undefined;
 
@@ -27,31 +27,6 @@ type RangeSelectorProps =
       setStart: (date: Date | undefined) => void;
       setEnd: (date: Date | undefined) => void;
     };
-
-function getWeekRange(date: Date) {
-  const minDate = getMinDate();
-  const maxDate = getMaxDate();
-  let startDay = new Date(date);
-  let endDay: Date;
-
-  startDay.setDate(date.getDate() - startDay.getDay());
-  if (startDay < minDate) {
-    startDay = new Date(minDate);
-  }
-
-  if (startDay.getTime() === minDate.getTime()) {
-    // データのある最初の週は7日周期にできないため、特別に終了日を設定
-    endDay = new Date(FIRST_WEEK_END_DATE);
-  } else {
-    endDay = new Date(startDay);
-    endDay.setDate(startDay.getDate() + 6);
-    // 最新データ日を超えないようにする
-    if (endDay > maxDate) {
-      endDay = new Date(maxDate);
-    }
-  }
-  return { from: startDay, to: endDay };
-}
 
 /**
  * 日付が選択可能な期間内かどうかを判定する
